fix(vue-query/queryOptions): accept the same options as useQuery
