+++ packages/vue-query/src/__tests__/queryOptions.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, it } from 'vitest'
 import { queryKey } from '@tanstack/query-test-utils'
 import { queryOptions } from '../queryOptions'
-import type { QueryOptions } from '../queryOptions'
+import type { UseQueryOptions } from '../queryOptions'
 
 describe('queryOptions', () => {
   it('should return the object received as a parameter without any modification.', () => {
-    const object: QueryOptions = {
+    const object: UseQueryOptions = {
       queryKey: queryKey(),
       queryFn: () => Promise.resolve(5),
     } as const
